requirements.txt
six
pyopenssl
django>=1.9,<1.10
django-model-utils
jsonfield
